+++ app/models/spree/authentication_method.rb
@@ -1,5 +1,9 @@
 class Spree::AuthenticationMethod < ActiveRecord::Base
-  validates :provider, :api_key, :api_secret, presence: true
+  def self.provider_options
+    SolidusSocial.configured_providers.map { |provider_name| [provider_name.split("_").first.camelize, provider_name] }
+  end
+
+  validates :provider, presence: true
 
   def self.active_authentication_methods?
     where(environment: ::Rails.env, active: true).exists?