+++ app/models/spree/social_configuration.rb
@@ -1,5 +1,6 @@
 module Spree
   class SocialConfiguration < Preferences::Configuration
+    attr_accessor :providers
     preference :path_prefix, :string, default: 'users'
   end
 end