+++ spec/models/spree/social_configuration_spec.rb
@@ -1,0 +1,4 @@
+RSpec.describe Spree::SocialConfiguration do
+  it { is_expected.to respond_to(:path_prefix) }
+  it { is_expected.to respond_to(:providers) }
+end